flask-api/requirements.txt
Flask
opencv-python
numpy
Pillow
tensorflow
tensorflow-hub
